requirements.txt
easyocr
pytesseract
pillow
transformers
torch
torchvision
numpy
python-dotenv